Refactored to have crossDeviceFlow and MasterFlow
