+++ src/components/Uploader/index.js
@@ -24,9 +24,9 @@ const UploadInstructions = ({error, method, side}) =>
 const UploadError = ({error}) =>
   error && <div className={`${style.text} ${style.error}`}>{`${error.message}. ${error.instruction}.`}</div>
 
-const UploaderPure = ({method, side, onImageSelected, error, ...{showMobileInstructions}}) =>
+const UploaderPure = ({method, side, onImageSelected, error, ...{startCrossDevice}}) =>
   <div>
-    { isDesktop ? <SwitchDevice {...{showMobileInstructions}} /> : '' }
+    { isDesktop ? <SwitchDevice {...{startCrossDevice}} /> : '' }
     <Dropzone
       onDrop={([ file ])=> {
         //removes a memory leak created by react-dropzone
